Replace ApplicationDateComponent with view helpers
