--- app/components/application_date_component.html.erb
@@ -1,12 +1,0 @@
-<%=
-  case type
-  when :submitted_at
-    submitted_at_date
-  when :respond_by
-    respond_by_date
-  when :edit_by
-    edit_by_date
-  when :days_remaining
-    formatted_days_remaining
-  end
-%>
